features/02-ai-content-analysis/requirements.txt
# AI Content Analysis Feature Requirements

# Core dependency for Gemini AI
google-generativeai

# For loading .env files
python-dotenv